Moved logging functionality to logging.py + updated documentation
